+++ R/separate-rows.R
@@ -4,8 +4,8 @@
 #' separates the values and places each one in its own row.
 #'
 #' @inheritSection gather Rules for selection
+#' @inheritParams fill
 #' @inheritParams gather
-#' @inheritParams separate
 #' @param sep Separator delimiting collapsed values.
 #' @export
 #' @examples