+++ tests/testthat/test-spread.R
@@ -24,9 +24,7 @@ test_that("convert turns strings into integers", {
 test_that("duplicate values for one key is an error", {
   df <- tibble(x = factor(c("a", "b", "b")), y = c(1, 2, 2), z = c(1, 2, 2))
   expect_error(spread(df, x, y),
-    "Duplicate identifiers for rows (2, 3)",
-    fixed = TRUE
-  )
+               "Keys are shared for 2 rows:")
 })
 
 test_that("factors are spread into columns (#35)", {
@@ -260,4 +258,23 @@ test_that("spread with fill replaces implicit missing values", {
   df <- tibble(key = factor("a", levels = c("a", "b")), value = 1)
   out <- spread(df, key, value, fill = 2, drop = FALSE)
   expect_equal(out, tibble(a = 1, b = 2))
+})
+
+test_that("ulevels preserves original factor levels", {
+  x_na_lev <- factor(c("a", NA), exclude = NULL)
+  expect_equal(levels(ulevels(x_na_lev)), c("a", NA))
+
+  x_na_lev_extra <- factor(c("a", NA), levels = c("a", "b", NA), exclude = NULL)
+  expect_equal(levels(ulevels(x_na_lev_extra)), c("a", "b", NA))
+
+  x_no_na_lev <- factor(c("a", NA))
+  expect_equal(levels(ulevels(x_no_na_lev)), "a")
+
+  x_no_na_lev_extra <- factor(c("a", NA), levels = c("a", "b"))
+  expect_equal(levels(ulevels(x_no_na_lev_extra)), c("a", "b"))
+})
+
+test_that("ulevels returns unique elements of a list for a list input", {
+  test_list <- list(a = 1:6, b = 1:6)
+  expect_equal(ulevels(test_list), unique(test_list))
 })