+++ NEWS.md
@@ -1,5 +1,15 @@
 # tidyr 0.8.2.9000
 
+* `unnest()` preserves column being unnested when input is zero-length (#483).
+
+* `unnest()` will run with named and unnamed list-columns of same length
+  (@hlendway, #460).
+
+* `crossing()` (and hence `expand()`) now works with list-columns 
+  (#446, @SamanthaToet).
+
+* `crossing()` preserves factor levels (#410).
+
 # tidyr 0.8.2
 
 * `separate()` now accepts `NA` as a column name in the `into` argument to